requirements.txt
Flask==2.3.3
python-telegram-bot==20.6
SQLAlchemy==2.0.21
psycopg2-binary==2.9.9
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
qrcode==7.4.2
Pillow==10.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
python-dateutil==2.8.2
openpyxl==3.1.2
